+++ routes/web.php
@@ -4,4 +4,6 @@ use Illuminate\Support\Facades\Route;
 
 Route::get('/', function () {
     return view('welcome');
-});
+})->name('dashboard');
+
+require __DIR__.'/navigation.php';